qcacmn: Use uint16 for peer ID
Peer ID can be 16bit wide, use unsigned short to avoid
overflow.

+++ dp/wifi3.0/dp_types.h
@@ -1578,7 +1578,7 @@ struct dp_pdev {
 	uint8_t dp_peer_based_pktlog;
 
 	/* Cached peer_id from htt_peer_details_tlv */
-	uint8_t fw_stats_peer_id;
+	uint16_t fw_stats_peer_id;
 
 	/* qdf_event for fw_peer_stats */
 	qdf_event_t fw_peer_stats_event;
